PythonServer/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp_retry==2.4.6
aiohttp_socks==0.7.1
deprecated==1.2.13
fastapi==0.78.0
numpy==1.17.4
Pillow==9.1.1
pydantic==1.9.1
requests==2.22.0
setuptools==45.2.0
